requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
msal==1.24.0
python-dotenv==1.0.0
requests==2.31.0
